Sakura blossom processing
The intitial data I found on the site was a large selection of
biological data. As I began to look beyond the samples I discovered that
there was a price associated with the past data (and realtime data). As
such I decided to see if there was any specific sakura data that was
free. I found some samples that were in an odd character encoding,
thankfully someone on stack exchange already ran across a similar issue
and the suggested encoding issue was quickly resolved. Now I just need
to start translating some of this.