Added batch translation function
I added a function that does the google translations in batches, and had
it re-instantiate the translator extension so as not to run afoul of Google
Translate's HTTP/429 error.

Everything seems to be working fine now, next is to actually find the
actual historical data. I thought this was it initially, but after
getting it translated, it's clearly not.