Started different dataset for Sakura data processing
It looks like the data from the Japan Meteorological Corporation
is just specification data. And I suspect it may only show up from
Feb to Apr. Which isn't super useful.

Anyway, it did lead me to finding the correct japanese terminology
to search for on google.co.jp. With that I was able to find the sakura
blossom data that the Japan Meteorological Agency has. And they have
data all the way back to the 50s. The biggest issue is that it's in
flat text format on their webpages. Processing that has been a bit of a
mess, but I think I have something figured out (finally). Basically
in the process of running it through regex101.com I noticed that
the data was in a fixed width format. With that I could then use pandas
fixed width format reader to read it into separate columns for me and
parse it from there.