Batch processing of the URLs
Started doing batch processing of the URLs and added in a some sleep
functionality for the processing so that the end site doesn't get
overloaded.

However, on viewing the reoconstructed data frame I saw that the first
set of data wasn't parsed correctly. This is because the columns
specifications were hardset because I wasn't able to get pandas column
inference to work correclty (and tab delimitation doesn't work either).

To solve this I added some test code below to work thorugh a more
dynamic way of dealing with the data (assuming an arbitrary number of
years surrounded by the same beginning and end columns). It's just about
working, but I need to further test it to ensure it's working properly.