Date fixes for dates from previous years (December)
In the process of fixing duplicate columns I noticed that some of the #
symbols weren't being parsed out correctly. Looking into this further I
discovered that there were some dates in December of 1990 that should
have been December of 1989. In the original source they were marked by
the # symbol which I hadn't been checking for yet.

I put in checks for the #, which mostly seems to be working. The only
stragglers are the #'s that got misalligned and put in the wrong column.
I need to put in additional data sanity checks.

TODO: Additional data integrity tests for Dates > October, but in the
current year.