Fixed Date processing and added translation
A bunch of issues with the date processing. The first one was a known
issue where Dates in December (marked with #) needed to be handled
correctly. I got that working and then moved on to fixing the duplicate
comment columns. I found one row that was breaking things, and it turned
out that row had a whole bunch of issues. In the process of fixing that
I increased the number of spaces in the dynamic column detection. That
fixed a bunch of issues.

I added translation. Mostly just using the code from my other noteobook,
although I had to modify it to use Japanese as a source language since
some of the city names got converted improperly because it detected the
language as Chinese (which is a valid mistake when working with Kanji).
Anyway, fixed that, and things look to be mostly working. I just need to
translate the final columns and figure out how to reorder things for a
nicer display (when I finally get to exporting this to a CSV file).