Initial commit of Temp notebook
This is against the original standard I implemented for the repository,
but I recently had to do a reinstall, so I'm doing a quick commit right
now to ensure I don't loose anything (since not everything is set back
up yet, other than the a basic anaconda + jupyter lab install).

Things are mostly documented in the notebook, but I made some progress
as far as finding a datafile that has a list of their site numbers. It
might be elsewhere too, but this is the first one I found. It's not
fully correct, in the POST request Kyoto is 's47759' and in the data
file the station number is 759, but that's close enough, and the last 3
digits match up for Osaka and Nara. The first two digits may change
based on region. I'll look into that next.