Hydrogen conversion and some POST testing
Tried out the atom editor with the hydrogen package to see how I liked
it compared to straight up jupyter lab.

In the process of trying that out started putting together a completely
unsuccesful POST request to the JMA site. The form data is the same as
what I got from chrome, so it "should" work, but it's throwing an
unhelfpul error page back at me.