Made a true error message display. Upped wait time
After getting a couple error messages displayed I tracked down
which particular element/selector was the actual error message.
That now displays instead of a generic error message.

I also made a quick function to download *all* of the data
based on what's from the smaster.index file that I got the
site conversions from. At the current wait time of 70s it
is going to take awhile to download, so I'm going to wait
until it's late night in Japan to run this fully.

I also am saving the data frames directly into CSVs and
keeping them in a dictionary. It's a bit redundant, but
it's what it is for now. I plan to re-write things a bit
so that it won't download a site that's already been fully
downloaded.