First pass at processing incomplete data
Reloaded all of the data back into a dictionary of dataframes. This is
so I can process what data I've already completely downloaded, and which
data I either need to redownload, or see if some of my assupmtions
regarding the naming conventions for the sites is incorrect.

Anyway also as part of reloading everything I also added some consecutive
data checks and number of data points to a dataframe so that I can
filter down to just what needs to be redownloaded. Or possibly a smaller
set of sites to go back and manually determine what nomenclature is
being used to identify their sites.

On the nomenclature: The site numbers should be correct; however, when I
was intitially investigating the data I noticed that large sites in a
prefecture followed the convention I was using (e.g. Osaka/Kyoto),
whereas smaller sites in the prefecture used a different nomenclature
for the prefix of the site id. I suspect that the remaining id
information from the master file should be fine, it's just a matter of
finding the correct prefix.