Added integration
Cleaned thigns up and added the code to actually do the integration to see where the maximum is. Also added in a half GCD to the end since there's usually
some space at the end to cast a filler Ruin III until the next Demi-Summon is up. Obviously lots of variance based on uptime and what not, but half a GCD
sounds about right based on my experience. (can probably fully verify later).