Sakura 2025 Updates: Depreciations and async translations
Mostly fixes for getting the code working again. Some (future)
depreciation changes for various pandas, etc features.

Changed the translation code to use async io. Which made the
code a whole lot simpler to read. It also allowed me to forgo
using my own rate limiting code for translations. It's still
there in case I need it in the future, but it's mostly not used
at this point.

Output CSV's uploaded to Kaggle.